+++ packages/ridb-react/src/index.tsx
@@ -3,7 +3,7 @@ import { RIDB,  WasmInternal } from "@trust0/ridb";
 import { BasePlugin, MigrationsParameter } from '@trust0/ridb-core';
 import { SchemaTypeRecord } from '@trust0/ridb-core';
 
-await WasmInternal();
+
 
 
 type DatabaseProps<T extends SchemaTypeRecord> = {
@@ -32,6 +32,9 @@ export function useDatabase<T extends SchemaTypeRecord>(): RIDB<T> {
 export function Database<T extends SchemaTypeRecord>({ children, ...props }: DatabaseComponentProps<T>) {
   const dbInit = props as DatabaseProps<T>;
   const db = useMemo(() => new RIDB<T>(dbInit), [props]);
+  useEffect(() => {
+    WasmInternal();
+  }, []);
   return (
     <DatabaseContext.Provider value={db}>
       {children}