+++ packages/ridb-core/CHANGELOG.md
@@ -1,3 +1,24 @@
+## 1.7.2 (2025-05-07)
+
+### 🩹 Fixes
+
+- lock issue ([6a5dea5](https://github.com/trust0-project/RIDB/commit/6a5dea5))
+- move to cjs modules ([d6c7608](https://github.com/trust0-project/RIDB/commit/d6c7608))
+
+### ❤️ Thank You
+
+- Javier Ribó
+
+## 1.7.1 (2025-05-07)
+
+### 🩹 Fixes
+
+- cleaning packages ([c6fba7e](https://github.com/trust0-project/RIDB/commit/c6fba7e))
+
+### ❤️ Thank You
+
+- Javier Ribó
+
 ## 1.7.0 (2025-05-07)
 
 ### 🚀 Features